# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Authn Platform API
  description: Enterprise Identity Engine REST API specification for authentication, session management, organization multi-tenancy,
    RBAC, SAML, OAuth2/OIDC, Webhooks, API key administration, and account recovery.
  version: 2.0.0
  contact:
    name: Authn Platform Engineering
    url: https://github.com/hanan-bhatti/authn-2.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
- url: http://localhost:8080
  description: Local Development Engine
- url: https://api.staging.authn.dev
  description: Staging Environment
- url: https://api.authn.dev
  description: Production Environment
paths:
...
  /v1/client/sessions:
    get:
      summary: List Authenticated User Active Sessions
      operationId: listSessions
      tags:
      - Session Management
      description: 'Returns the calling user''s active sessions, flagging which session processed this request.

        '
      security:
      - PublishableKey: []
        BearerAuth: []
      responses:
        '200':
          description: List of active sessions
          content:
            application/json:
              schema:
                type: object
                required:
                - sessions
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionResponse'
        '401':
          description: Unauthorized request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error: 'session authentication required: missing or invalid access token'
                code: unauthorized
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/client/sessions/revoke:
    post:
      summary: Revoke Specific User Session
      operationId: revokeSession
      tags:
      - Session Management
      description: 'Revokes a single active session belonging to the caller specified in the request body.

        '
      security:
      - PublishableKey: []
        BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeSessionRequest'
      responses:
        '200':
          description: Session revoked successfully
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                - session_id
                properties:
                  message:
                    type: string
                    example: session revoked
                  session_id:
                    type: string
                    example: sess_01H1234567890ABCDEF
        '400':
          description: Missing session_id parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error: session_id is required
                code: missing_parameter
        '401':
          description: Unauthorized request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error: 'session authentication required: missing or invalid access token'
                code: unauthorized
        '403':
          description: Session belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error: session does not belong to the calling user
                code: forbidden
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error: session not found
                code: not_found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/client/sessions/revoke-others:
    post:
      summary: Revoke Other User Sessions
      operationId: revokeOtherSessions
      tags:
      - Session Management
      description: 'Revokes all active sessions belonging to the caller except the session executing this request.

        '
      security:
      - PublishableKey: []
        BearerAuth: []
      responses:
        '200':
          description: Other sessions revoked successfully
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                - count
                properties:
                  message:
                    type: string
                    example: all other sessions revoked
                  count:
                    type: integer
                    example: 2
        '401':
          description: Unauthorized request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error: 'session authentication required: missing or invalid access token'
                code: unauthorized
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/client/sessions/revoke-all:
    post:
      summary: Revoke All User Sessions
      operationId: revokeAllSessions
      tags:
      - Session Management
      description: 'Revokes every session belonging to the caller, including the current active session.

        '
      security:
      - PublishableKey: []
        BearerAuth: []
      responses:
        '200':
          description: All sessions revoked successfully
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                - count
                properties:
                  message:
                    type: string
                    example: all sessions revoked
                  count:
                    type: integer
                    example: 3
        '401':
          description: Unauthorized request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error: 'session authentication required: missing or invalid access token'
                code: unauthorized
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/client/auth/refresh:
    post:
      summary: Refresh Access Token
      operationId: refreshTokens
      tags:
      - Session Management
      description: 'Exchanges a refresh token (supplied in JSON request body or `authn_refresh_token` HTTP cookie) for a new
        access token and rotated refresh token.

        '
      security:
      - PublishableKey: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokensRequest'
      responses:
        '200':
          description: Token pair refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionTokenPairResponse'
        '400':
          description: Missing refresh token parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              example:
                error: refresh_token required
                code: missing_parameter
        '401':
          description: Refresh token invalid, expired, revoked, or compromised
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
              examples:
                session_compromised:
                  summary: Token reuse detected / session compromised
                  value:
                    error: session revoked due to token reuse detection
                    code: session_compromised
                session_revoked:
                  summary: Session explicitly revoked
                  value:
                    error: session has been revoked
                    code: session_revoked
                session_expired:
                  summary: Session expired
                  value:
                    error: session has expired
                    code: session_expired
                invalid_token:
                  summary: Unknown or invalid refresh token
                  value:
                    error: invalid or expired refresh token
                    code: invalid_token
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/admin/users/{user_id}/sessions:
    get:
      summary: Admin List User Active Sessions
      operationId: adminListUserSessions
      tags:
      - Admin Session Management
      description: 'Admin endpoint to list all active sessions for a target user. Requires secret key or tenant admin access.

        '
      parameters:
      - name: user_id
        in: path
        required: true
        description: Target user ID whose sessions to list
        schema:
          type: string
      security:
      - SecretKey: []
      - BearerAuth: []
      responses:
        '200':
          description: Target user's active sessions
          content:
            application/json:
              schema:
                type: object
                required:
                - sessions
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/SessionResponse'
        '401':
          description: Unauthorized request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden / Admin permissions required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /v1/admin/users/{user_id}/sessions/revoke-all:
    post:
      summary: Admin Revoke All Sessions for User
      operationId: adminRevokeAllUserSessions
      tags:
      - Admin Session Management
      description: 'Admin endpoint to forcibly revoke all active sessions for a target user. Requires secret key or tenant
        admin access.

        '
      parameters:
      - name: user_id
        in: path
        required: true
        description: Target user ID whose sessions to revoke
        schema:
          type: string
      security:
      - SecretKey: []
      - BearerAuth: []
      responses:
        '200':
          description: All sessions revoked for target user
          content:
            application/json:
              schema:
                type: object
                required:
                - message
                - user_id
                - count
                properties:
                  message:
                    type: string
                    example: all sessions revoked for user
                  user_id:
                    type: string
                    example: usr_01H1234567890ABCDEF
                  count:
                    type: integer
                    example: 3
        '401':
          description: Unauthorized request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden / Admin permissions required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  securitySchemes:
    PublishableKey:
      type: apiKey
      name: X-Authn-Publishable-Key
      in: header
      description: Client publishable key identifying application context
    SecretKey:
      type: apiKey
      name: X-Authn-Secret-Key
      in: header
      description: Backend secret key for tenant admin API calls
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Client user or console admin session JWT
  schemas:
...
    ErrorEnvelope:
      type: object
      required:
      - error
      - code
      properties:
        error:
          type: string
          description: Human-readable error message (never discloses internal implementation details)
          example: invalid or expired verification token
        code:
          type: string
          description: Stable machine-readable error identifier
          example: invalid_token
...
    DeviceInfo:
      type: object
      required:
      - browser
      - os
      - device
      - label
      properties:
        browser:
          type: string
          example: Chrome
        os:
          type: string
          example: macOS
        device:
          type: string
          example: Desktop
        label:
          type: string
          example: Chrome on macOS
    SessionResponse:
      type: object
      required:
      - id
      - device
      - ip_address
      - location
      - created_at
      - is_current
      properties:
        id:
          type: string
          example: sess_01H1234567890ABCDEF
        device:
          $ref: '#/components/schemas/DeviceInfo'
        ip_address:
          type: string
          example: 192.0.2.1
        location:
          type: string
          example: Unknown Location
        last_active_at:
          type: string
          format: date-time
          example: '2026-08-10T10:00:00Z'
        created_at:
          type: string
          format: date-time
          example: '2026-08-01T12:00:00Z'
        is_current:
          type: boolean
          example: true
    RevokeSessionRequest:
      type: object
      required:
      - session_id
      properties:
        session_id:
          type: string
          example: sess_01H1234567890ABCDEF
    RefreshTokensRequest:
      type: object
      properties:
        refresh_token:
          type: string
          description: Opaque refresh token (can also be supplied via authn_refresh_token cookie)
          example: rt_01H1234567890ABCDEF
    SessionTokenPairResponse:
      type: object
      required:
      - access_token
      - token_type
      - expires_in
      - session_id
      properties:
        access_token:
          type: string
          description: Newly issued JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: Rotated refresh token (omitted / empty string on grace window replay)
          example: rt_9876543210FEDCBA
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          description: Access token TTL in seconds
          example: 3600
        session_id:
          type: string
          example: sess_01H1234567890ABCDEF